# docs/swagger.yaml
basePath: /v1
definitions:
  main.ExecRegisterPayload:
    properties:
      email:
        type: string
      first_name:
        maxLength: 72
        type: string
      last_name:
        maxLength: 72
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - manager
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - role
    type: object
...
  main.StudentRegisterPayload:
    properties:
      address:
        type: string
      birth_date:
        type: string
      class:
        type: string
      email:
        type: string
      first_name:
        maxLength: 72
        type: string
      last_name:
        maxLength: 72
        type: string
      parent_name:
        type: string
      parent_phone_number:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      phone_number:
        type: string
      teacher_id:
        type: integer
    required:
    - address
    - birth_date
    - class
    - email
    - first_name
    - last_name
    - parent_name
    - parent_phone_number
    - password
    - teacher_id
    type: object
  main.TeacherRegisterPayload:
    properties:
      email:
        type: string
      first_name:
        maxLength: 72
        type: string
      hire_date:
        type: string
      last_name:
        maxLength: 72
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      phone_number:
        type: string
      subject:
        maxLength: 128
        type: string
    required:
    - email
    - first_name
    - hire_date
    - last_name
    - password
    - phone_number
    - subject
    type: object
...
  main.UpdateStudentPayload:
    properties:
      address:
        maxLength: 256
        type: string
      birth_date:
        type: string
      class:
        maxLength: 16
        type: string
      email:
        type: string
      first_name:
        maxLength: 72
        type: string
      last_name:
        maxLength: 72
        type: string
      parent_name:
        maxLength: 128
        type: string
      parent_phone_number:
        type: string
      phone_number:
        type: string
      teacher_id:
        type: integer
    type: object
...
  store.Student:
    properties:
      address:
        type: string
      birth_date:
        type: string
      class:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      parent_name:
        type: string
      parent_phone_number:
        type: string
      phone_number:
        type: string
      teacher_id:
        type: integer
      updated_at:
        type: string
    type: object
  store.Teacher:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      hire_date:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone_number:
        type: string
      subject:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for ClassNama, a school management platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ClassNama
paths:
...
  /execs/register:
    post:
      consumes:
      - application/json
      description: Only Execs with manager/admin roles can create new Execs
      parameters:
      - description: Exec registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.ExecRegisterPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Returns the created Exec and JWT token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new Exec
      tags:
      - Execs
  /students:
    get:
      operationId: getStudents
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Student'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get all students
      tags:
      - Students
    post:
      consumes:
      - application/json
      description: Only Execs with manager/admin roles can create new Students
      parameters:
      - description: Student registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.StudentRegisterPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Returns the created Student
          schema:
            $ref: '#/definitions/store.Student'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new Student
      tags:
      - Students
  /students/{studentID}:
    delete:
      operationId: deleteStudent
      parameters:
      - description: student ID
        in: path
        name: studentID
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete a student
      tags:
      - Students
    get:
      operationId: getstudent
      parameters:
      - description: student ID
        in: path
        name: studentID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Student'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get a student by ID
      tags:
      - Students
    patch:
      consumes:
      - application/json
      operationId: updateStudent
      parameters:
      - description: student ID
        in: path
        name: studentID
        required: true
        type: integer
      - description: student update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateStudentPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Student'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "409":
          description: Conflict
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update a student
      tags:
      - Students
...
  /teachers:
    get:
      operationId: getTeachers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Teacher'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get all teachers
      tags:
      - Teachers
    post:
      consumes:
      - application/json
      description: Only Execs with manager/admin roles can create new Teachers
      parameters:
      - description: Teacher registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.TeacherRegisterPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Returns the created Teacher
          schema:
            $ref: '#/definitions/store.Teacher'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new Teacher
      tags:
      - Teachers
...
  /teachers/{teacherID}/students:
    get:
      consumes:
      - application/json
      description: Returns a list of all students assigned to a specific teacher
      operationId: getStudentsByTeacher
      parameters:
      - description: Teacher ID
        in: path
        name: teacherID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of students
          schema:
            items:
              $ref: '#/definitions/store.Student'
            type: array
        "400":
          description: Bad request
          schema: {}
        "404":
          description: Teacher not found / no students
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get students of a teacher
      tags:
      - Students
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"